feat(basecamp): add `basecamp doctor` + `launch` variant pre-flight
Per PR #75 review finding 6 + follow-up direction: basecamp-specific
health checks live under `basecamp doctor` (not merged into top-level
`logos-scaffold doctor`) to keep basecamp commands self-contained
during this PR. A follow-up can merge rows once the core feature lands.

New `logos-scaffold basecamp doctor` subcommand:

- **Captured modules summary** — one row per captured source (both
  `project_sources` and `dependencies`) showing:
  - Flake ref verbatim.
  - Parsed `(tag <name>)` or `(commit <12-hex>)` annotation for
    github refs (non-hex middle segment → tag; 7–40-char hex → commit).
  - API header paths (`*.h` / `*.hpp`) found under alice's installed
    `modules/<name>/`, `include/`, or `interfaces/` subdir — helps
    devs locate IPC headers without digging through `.lgx` extracts.

- **Manifest variant check** — walks each seeded profile's
  `modules/*/manifest.json`, flags any whose `main` object lacks the
  current-platform `-dev` variant key (`linux-amd64-dev`, etc.).
  Catches the freeze-on-first-click footgun vpavlin surfaced.

- **Drift check** — runs the `basecamp modules` auto-discovery
  algorithm as a dry-run and diffs against `state.sources`. Warns on
  discovered-but-uncaptured (new sub-flake / new manifest dep) and
  captured-but-no-longer-discoverable (deleted sub-flake / removed dep).

- Supports `--json` for machine-readable output (same `DoctorReport`
  shape as top-level doctor).

Also lands `launch` variant pre-flight:

- Before exec, walks `<profile>/modules/*/manifest.json` with the
  same `check_manifest_variants` helper.
- Prints a single-line summary: `launch: profile alice has N
  module(s); all linux-amd64-dev variants present ✓` on success, or
  `… N missing linux-amd64-dev variant (<names>); run `logos-scaffold
  doctor` for details — plugins will hang on click.` on mismatch.
- Non-blocking: warning surfaces, launch proceeds.

Refactor: top-level `doctor.rs` factors `finalize_report(rows)` and
`print_report(&report)` so `basecamp doctor` reuses the same
formatting + summary + next-steps logic.

Basecamp-specific helpers that were transiently in doctor.rs (drift,
variant check, github-ref parser, module-name inference, header walk)
now live in `commands/basecamp.rs` where they belong. Five unit tests
for the ref / name / header helpers round-trip there too.

Finding 6's upstream suggestion (emit both `<plat>` and `<plat>-dev`
keys in `logos-module-builder` packaging) will be filed separately.
